Fixed note scroll speed defaulting to 0.5x for newly registered profiles. Introduced several new interfaces, and refactored code to cleanup profiles.ts
