+++ gitadora@asphyxia/data/mdb/.gitignore
@@ -8,3 +8,4 @@ nt.json
 hv.json
 custom.xml
 custom.json
+blacklist.txt
