--- gitadora@asphyxia/models/playerstickerresponse.ts
@@ -1,8 +1,0 @@
-export interface PlayerStickerResponse {
-    id: KITEM<'s32'>,
-    pos_x: KITEM<'float'> ,
-    pos_y: KITEM<'float'>,
-    scale_x: KITEM<'float'> ,
-    scale_y: KITEM<'float'>,
-    rotate: KITEM<'float'>
-}
