+++ gitadora@asphyxia/models/profile.ts
@@ -1,3 +1,4 @@
+import { PLUGIN_VER } from "../const";
 import { SecretMusicEntry } from "./secretmusicentry";
 
 export interface Profile {
@@ -60,3 +61,66 @@ export interface Profile {
     music: SecretMusicEntry[];    
   }
 }
+
+export function getDefaultProfile (game: 'gf' | 'dm', version: string, id: number): Profile {
+  return {
+    collection: 'profile',
+    pluginVer: PLUGIN_VER,
+
+    game,
+    version,
+    id,
+
+    play: 0,
+    playtime: 0,
+    playterm: 0,
+    session_cnt: 0,
+    extra_stage: 0,
+    extra_play: 0,
+    extra_clear: 0,
+    encore_play: 0,
+    encore_clear: 0,
+    pencore_play: 0,
+    pencore_clear: 0,
+    max_clear_diff: 0,
+    max_full_diff: 0,
+    max_exce_diff: 0,
+    clear_num: 0,
+    full_num: 0,
+    exce_num: 0,
+    no_num: 0,
+    e_num: 0,
+    d_num: 0,
+    c_num: 0,
+    b_num: 0,
+    a_num: 0,
+    s_num: 0,
+    ss_num: 0,
+    last_category: 0,
+    last_musicid: -1,
+    last_seq: 0,
+    disp_level: 0,
+    progress: 0,
+    disp_state: 0,
+    skill: 0,
+    all_skill: 0,
+    extra_gauge: 0,
+    encore_gauge: 0,
+    encore_cnt: 0,
+    encore_success: 0,
+    unlock_point: 0,
+    max_skill: 0,
+    max_all_skill: 0,
+    clear_diff: 0,
+    full_diff: 0,
+    exce_diff: 0,
+    clear_music_num: 0,
+    full_music_num: 0,
+    exce_music_num: 0,
+    clear_seq_num: 0,
+    classic_all_skill: 0,
+    secretmusic: {
+      music: []     
+    }
+  }
+};
