--- gitadora@asphyxia/models/secretmusicresponse.ts
@@ -1,5 +1,0 @@
-export interface SecretMusicResponse {
-    musicid: KITEM<'s32'>;
-    seq: KITEM<'u16'>;
-    kind: KITEM<'s32'>;
-}
